meta-balena-raspberrypi/linux: Add patch to fix overlayfs regression
The kernel version we pull in from meta-raspberry seems to hang in between
two commits that break and subsequently "unbreak" how the kernel deals with
overlapping parent layers in overlayfs.

> User is not supposed to specify the same directory for upper and
> lower dirs or for different lower layers and user is not supposed to
> specify directories that are descendants of each other for overlay
> layers

This is something that docker relies on heavily and we've also seen this fail
more frequently when deltas are enabled.

The solution generally would be to have:
https://github.com/raspberrypi/linux/commit/9c13e770cca978b8d5c870f2270cdf3a10c63817

Until we update meta-raspberry to include this, this fixes the problem.

Change-type: patch
Changelog-entry: Add a patch for linux 4.19 that fixes a regression in overlayfs
Signed-off-by: Robert Günzler <robertg@balena.io>